requirements.txt
# Requirements for German translation script
openai>=1.99.0
python-dotenv>=1.0.0